(tests) Create dedicated file for check pwd exists
to have the test binary is built and run seperately.

This is necessary as changing the working directory
within the tests likely breaks other tests executed
in parallel by changing the working directory.

Also trim down the test logic a bit. "--no-cache" is
redundant.

Use imports instead of qualifiers for readability.
